+++ js/main.js
@@ -2,3 +2,5 @@ import './slider.js';
 import './user-validation.js';
 import './map.js';
 import './data.js';
+import './api.js';
+
